Move InteractableDrawing custom painter to a separate file
