+++ lib/src/deriv_chart/interactive_layer/interactive_layer.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tool_config.dart';
 import 'package:deriv_chart/src/add_ons/repository.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/edge_point.dart';
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/models/animation_info.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/gestures/gesture_manager.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/x_axis/x_axis_model.dart';
 import 'package:deriv_chart/src/models/chart_config.dart';
@@ -14,8 +13,8 @@ import 'package:provider/provider.dart';
 import '../chart/data_visualization/chart_data.dart';
 import '../chart/data_visualization/chart_series/data_series.dart';
 import '../chart/data_visualization/drawing_tools/ray/ray_line_drawing.dart';
-import '../chart/y_axis/y_axis_config.dart';
 import '../drawing_tool_chart/drawing_tools.dart';
+import 'interactable_drawing_custom_painter.dart';
 
 /// Interactive layer of the chart package where elements can be drawn and can
 /// be interacted with.
@@ -259,7 +258,7 @@ class _InteractiveLayerState extends State<InteractiveLayer> {
         children: [
           ..._interactableDrawings
               .map((e) => CustomPaint(
-                    foregroundPainter: _DrawingPainter(
+                    foregroundPainter: InteractableDrawingCustomPainter(
                       drawing: e,
                       series: widget.series,
                       theme: context.watch<ChartTheme>(),
@@ -276,69 +275,5 @@ class _InteractiveLayerState extends State<InteractiveLayer> {
         ],
       ),
     );
-  }
-}
-
-/// A callback which calling it should return if the [drawing] is selected.
-typedef IsDrawingSelected = bool Function(InteractableDrawing drawing);
-
-class _DrawingPainter extends CustomPainter {
-  _DrawingPainter({
-    required this.drawing,
-    required this.series,
-    required this.theme,
-    required this.chartConfig,
-    required this.epochFromX,
-    required this.epochToX,
-    required this.quoteToY,
-    required this.quoteFromY,
-    required this.isSelected,
-  });
-
-  final InteractableDrawing drawing;
-  final DataSeries<Tick> series;
-  final ChartTheme theme;
-  final ChartConfig chartConfig;
-  final int Function(double x) epochFromX;
-  final double Function(int x) epochToX;
-  final double Function(double y) quoteToY;
-
-  double Function(double) quoteFromY;
-
-  // final Function() onDrawingToolClicked;
-
-  /// Returns `true` if the drawing tool is selected.
-  final bool Function(InteractableDrawing) isSelected;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    YAxisConfig.instance.yAxisClipping(canvas, size, () {
-      drawing.paint(
-        canvas,
-        size,
-        epochToX,
-        quoteToY,
-        const AnimationInfo(),
-        isSelected,
-      );
-      // TODO(NA): Paint the [drawing]
-    });
-  }
-
-  @override
-  bool shouldRepaint(_DrawingPainter oldDelegate) =>
-      // TODO(NA): Return true/false based on the [drawing] state
-      true;
-
-  @override
-  bool shouldRebuildSemantics(_DrawingPainter oldDelegate) => false;
-
-  @override
-  bool hitTest(Offset position) {
-    if (drawing.hitTest(position, epochToX, quoteToY)) {
-      // onDrawingToolClicked();
-      return true;
-    }
-    return false;
   }
 }
